requirements.txt
kafka-python==2.0.2
boto3==1.26.0
opencv-python==4.5.3.56    # If needed for OAK-D
numpy==1.23.0              # For data processing
